@StephanTLavavej 2025-03-06
Thanks! :heart_eyes_cat: I pushed a conflict-free merge with `main` and very minor nitpicks.

5950X speedups look good:

Benchmark                              |  Before |   After | Speedup
---------------------------------------|--------:|--------:|--------
`bm<char...NotFirstOne>/8021/3056`     |  646 ns | 50.4 ns | 12.82
`bm<char...NotFirstOne>/63/62`         | 16.9 ns | 4.86 ns | 3.48
`bm<char...NotFirstOne>/31/30`         | 9.39 ns | 8.89 ns | 1.06
`bm<char...NotFirstOne>/15/14`         | 5.23 ns | 8.15 ns | 0.64
`bm<char...NotFirstOne>/7/6`           | 3.25 ns | 5.78 ns | 0.56
`bm<char...NotLastOne>/8021/3056`      |  645 ns | 55.2 ns | 11.68
`bm<char...NotLastOne>/63/62`          | 15.2 ns | 5.74 ns | 2.65
`bm<char...NotLastOne>/31/30`          | 8.74 ns | 9.83 ns | 0.89
`bm<char...NotLastOne>/15/14`          | 4.72 ns | 9.19 ns | 0.51
`bm<char...NotLastOne>/7/6`            | 2.97 ns | 6.34 ns | 0.47
`bm<wchar_t...NotFirstOne>/8021/3056`  |  647 ns | 90.8 ns | 7.13
`bm<wchar_t...NotFirstOne>/63/62`      | 16.2 ns | 5.31 ns | 3.05
`bm<wchar_t...NotFirstOne>/31/30`      | 9.49 ns | 4.87 ns | 1.95
`bm<wchar_t...NotFirstOne>/15/14`      | 5.22 ns | 6.31 ns | 0.83
`bm<wchar_t...NotFirstOne>/7/6`        | 3.38 ns | 5.98 ns | 0.57
`bm<wchar_t...NotLastOne>/8021/3056`   |  645 ns | 95.5 ns | 6.75
`bm<wchar_t...NotLastOne>/63/62`       | 15.1 ns | 5.55 ns | 2.72
`bm<wchar_t...NotLastOne>/31/30`       | 8.37 ns | 5.14 ns | 1.63
`bm<wchar_t...NotLastOne>/15/14`       | 4.72 ns | 6.57 ns | 0.72
`bm<wchar_t...NotLastOne>/7/6`         | 2.75 ns | 6.10 ns | 0.45
`bm<char32_t...NotFirstOne>/8021/3056` |  647 ns |  177 ns | 3.66
`bm<char32_t...NotFirstOne>/63/62`     | 16.6 ns | 6.55 ns | 2.53
`bm<char32_t...NotFirstOne>/31/30`     | 9.85 ns | 4.86 ns | 2.03
`bm<char32_t...NotFirstOne>/15/14`     | 5.84 ns | 4.05 ns | 1.44
`bm<char32_t...NotFirstOne>/7/6`       | 3.60 ns | 4.67 ns | 0.77
`bm<char32_t...NotLastOne>/8021/3056`  |  645 ns |  171 ns | 3.77
`bm<char32_t...NotLastOne>/63/62`      | 15.0 ns | 6.19 ns | 2.42
`bm<char32_t...NotLastOne>/31/30`      | 8.35 ns | 5.13 ns | 1.63
`bm<char32_t...NotLastOne>/15/14`      | 4.46 ns | 4.70 ns | 0.95
`bm<char32_t...NotLastOne>/7/6`        | 2.76 ns | 5.32 ns | 0.52